resolves #110

+++ src/main/java/com/github/kiulian/downloader/model/playlist/PlaylistDetails.java
@@ -7,9 +7,9 @@ public class PlaylistDetails {
     private String title;
     private String author;
     private int videoCount;
-    private int viewCount;
+    private long viewCount;
 
-    public PlaylistDetails(String playlistId, String title, String author, int videoCount, int viewCount) {
+    public PlaylistDetails(String playlistId, String title, String author, int videoCount, long viewCount) {
         super();
         this.playlistId = playlistId;
         this.title = title;
@@ -34,7 +34,7 @@ public class PlaylistDetails {
         return videoCount;
     }
 
-    public int viewCount() {
+    public long viewCount() {
         return viewCount;
     }
 }
